refactor: extract customer, file, manifest, S3, cluster, monitoring routes from workset_api.py
Created 6 new router modules following existing create_X_router(deps) pattern:
- daylib/routes/customers.py: Customer CRUD (5 endpoints)
- daylib/routes/files.py: File management (6 endpoints)
- daylib/routes/manifests.py: Manifest operations (4 endpoints)
- daylib/routes/s3.py: S3 utilities (5 endpoints)
- daylib/routes/clusters.py: Cluster management (2 endpoints)
- daylib/routes/monitoring.py: Admin monitoring (1 endpoint)

workset_api.py reduced from 2,287 to 770 lines (thin app factory).
Updated test mocks to reference new module locations.
All 746 tests pass.

Agent-Id: agent-bfd43cb2-09b9-4b78-b001-f3f31e2d2c76
