+++ daylib/workset_api.py
@@ -1,55 +1,35 @@
 """FastAPI web interface for workset monitoring and management.
 
-Provides REST API and web dashboard for workset operations.
-
-This module uses shared Pydantic models from daylib.routes.dependencies.
+Thin app factory that assembles routers and middleware.
+Route handlers live in daylib/routes/*.py modules.
 """
 
 from __future__ import annotations
 
-import gzip
-import io
 import logging
-import re
-import tarfile
 import uuid
-import zipfile
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 from pathlib import Path
-from typing import Any, DefaultDict, Dict, List, Optional, Tuple
+from typing import Any, Dict, Optional, Tuple
 
-import boto3
 import yaml  # type: ignore[import-untyped]
 
-from fastapi import Body, Depends, FastAPI, File, Form, HTTPException, Query, Request, UploadFile, status
+from fastapi import Body, Depends, FastAPI, HTTPException, Query, Request, status
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import Response
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
 from starlette.middleware.sessions import SessionMiddleware
 
 from daylib.config import Settings, get_settings
-from daylib.security import sanitize_for_log
-from daylib.workset_state_db import WorksetPriority, WorksetState, WorksetStateDB
+from daylib.workset_state_db import WorksetStateDB
 from daylib.workset_scheduler import WorksetScheduler
 
-# Import shared Pydantic models from routes module
+# Import shared models and utilities used in remaining inline endpoints
 from daylib.routes.dependencies import (
-    WorksetCreate,
-    WorksetResponse,
-    WorksetStateUpdate,
-    QueueStats,
-    SchedulingStats,
-    CustomerCreate,
-    CustomerUpdate,
-    CustomerResponse,
     WorksetValidationResponse,
     WorkYamlGenerateRequest,
-    format_file_size as _format_file_size,
-    get_file_icon as _get_file_icon,
     calculate_cost_with_efficiency as _calculate_cost_with_efficiency,
     convert_customer_for_template as _convert_customer_for_template,
-    verify_workset_ownership,
 )
 
 # Optional imports for authentication
@@ -63,15 +43,6 @@ except ImportError:
 
 from daylib.workset_customer import CustomerManager, CustomerConfig
 from daylib.workset_validation import WorksetValidator
-
-# Pipeline status monitoring
-try:
-    from daylib.pipeline_status import PipelineStatusFetcher, PipelineStatus
-    PIPELINE_STATUS_AVAILABLE = True
-except ImportError:
-    PIPELINE_STATUS_AVAILABLE = False
-    PipelineStatusFetcher = None  # type: ignore[misc, assignment]
-    PipelineStatus = None  # type: ignore[misc, assignment]
 
 # Optional integration layer import
 try:
@@ -107,24 +78,13 @@ except ImportError:
 
 # Manifest storage imports
 try:
-    from daylib.manifest_registry import ManifestRegistry, ManifestTooLargeError
+    from daylib.manifest_registry import ManifestRegistry
     MANIFEST_STORAGE_AVAILABLE = True
 except ImportError:
     MANIFEST_STORAGE_AVAILABLE = False
     ManifestRegistry = None  # type: ignore[misc, assignment]
-    ManifestTooLargeError = None  # type: ignore[misc, assignment]
 
 LOGGER = logging.getLogger("daylily.workset_api")
-
-
-# Note: Pydantic models and utility functions are now imported from daylib.routes.dependencies
-# See the imports at the top of this file for:
-# - WorksetCreate, WorksetResponse, WorksetStateUpdate, QueueStats, SchedulingStats
-# - CustomerCreate, CustomerResponse, WorksetValidationResponse, WorkYamlGenerateRequest
-# - ChangePasswordRequest, APITokenCreateRequest
-# - PortalFileAutoRegisterRequest, PortalFileAutoRegisterResponse
-# - _format_file_size, _get_file_icon, _calculate_cost_with_efficiency, _convert_customer_for_template
-# - verify_workset_ownership
 
 
 def create_app(
@@ -467,184 +427,64 @@ def create_app(
             "timestamp": datetime.now(timezone.utc).isoformat().replace("+00:00", "Z"),
         }
     
-    @app.post("/worksets", response_model=WorksetResponse, status_code=status.HTTP_201_CREATED, tags=["worksets"])
-    async def create_workset(workset: WorksetCreate):
-        """Register a new workset."""
-        try:
-            success = state_db.register_workset(
-                workset_id=workset.workset_id,
-                bucket=workset.bucket,
-                prefix=workset.prefix,
-                priority=workset.priority,
-                workset_type=workset.workset_type,
-                metadata=workset.metadata,
-                customer_id=workset.customer_id,
-                preferred_cluster=workset.preferred_cluster,
-            )
-        except ValueError as e:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=str(e),
-            )
+    # ========== Wire up extracted routers ==========
 
-        if not success:
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail=f"Workset {workset.workset_id} already exists",
-            )
+    # Workset CRUD + queue/scheduler/next (routes/worksets.py)
+    from daylib.routes.worksets import create_worksets_router
+    worksets_router = create_worksets_router(state_db=state_db, scheduler=scheduler)
+    app.include_router(worksets_router)
+    LOGGER.info("Workset routes registered via create_worksets_router")
 
-        # Retrieve the created workset
-        created = state_db.get_workset(workset.workset_id)
-        if not created:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail="Failed to retrieve created workset",
-            )
+    # Monitoring/admin routes (routes/monitoring.py)
+    from daylib.routes.monitoring import create_monitoring_router, MonitoringDependencies
+    mon_deps = MonitoringDependencies(state_db=state_db, settings=settings)
+    mon_router = create_monitoring_router(mon_deps)
+    app.include_router(mon_router)
+    LOGGER.info("Monitoring routes registered via create_monitoring_router")
 
-        return WorksetResponse(**created)
-    
-    @app.get("/worksets/{workset_id}", response_model=WorksetResponse, tags=["worksets"])
-    async def get_workset(workset_id: str):
-        """Get workset details."""
-        workset = state_db.get_workset(workset_id)
-        if not workset:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Workset {workset_id} not found",
-            )
-        
-        return WorksetResponse(**workset)
-    
-    @app.get("/worksets", response_model=List[WorksetResponse], tags=["worksets"])
-    async def list_worksets(
-        state: Optional[WorksetState] = Query(None, description="Filter by state"),
-        priority: Optional[WorksetPriority] = Query(None, description="Filter by priority"),
-        customer_id: Optional[str] = Query(None, description="Filter by customer ownership"),
-        limit: int = Query(100, ge=1, le=1000, description="Maximum results"),
-    ):
-        """List worksets with optional filters.
+    # S3 utility routes (routes/s3.py)
+    from daylib.routes.s3 import create_s3_router, S3Dependencies
+    s3_deps = S3Dependencies(settings=settings, get_current_user=get_current_user)
+    s3_router = create_s3_router(s3_deps)
+    app.include_router(s3_router)
+    LOGGER.info("S3 routes registered via create_s3_router")
 
-        If customer_id is provided, only returns worksets owned by that customer.
-        """
-        # Fetch more if filtering by customer to account for non-matching worksets
-        fetch_limit = limit * 5 if customer_id else limit
+    # Cluster management routes (routes/clusters.py)
+    from daylib.routes.clusters import create_clusters_router, ClusterDependencies
+    cluster_deps = ClusterDependencies(settings=settings, get_current_user=get_current_user)
+    cluster_router = create_clusters_router(cluster_deps)
+    app.include_router(cluster_router)
+    LOGGER.info("Cluster routes registered via create_clusters_router")
 
-        if state:
-            worksets = state_db.list_worksets_by_state(state, priority=priority, limit=fetch_limit)
-        else:
-            # Get all states
-            worksets = []
-            for ws_state in WorksetState:
-                batch = state_db.list_worksets_by_state(ws_state, priority=priority, limit=fetch_limit)
-                worksets.extend(batch)
-                if len(worksets) >= fetch_limit:
-                    break
-            worksets = worksets[:fetch_limit]
-
-        # SECURITY: If customer_id provided, filter to only that customer's worksets
-        if customer_id:
-            worksets = [
-                w for w in worksets
-                if verify_workset_ownership(w, customer_id)
-            ]
-
-        return [WorksetResponse(**w) for w in worksets[:limit]]
-
-    @app.put("/worksets/{workset_id}/state", response_model=WorksetResponse, tags=["worksets"])
-    async def update_workset_state(workset_id: str, update: WorksetStateUpdate):
-        """Update workset state."""
-        # Verify workset exists
-        workset = state_db.get_workset(workset_id)
-        if not workset:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Workset {workset_id} not found",
-            )
-
-        state_db.update_state(
-            workset_id=workset_id,
-            new_state=update.state,
-            reason=update.reason,
-            error_details=update.error_details,
-            cluster_name=update.cluster_name,
-            metrics=update.metrics,
+    # Customer routes (routes/customers.py) - only if customer_manager available
+    if customer_manager:
+        from daylib.routes.customers import create_customers_router, CustomerDependencies
+        cust_deps = CustomerDependencies(
+            customer_manager=customer_manager,
+            get_current_user=get_current_user,
         )
+        cust_router = create_customers_router(cust_deps)
+        app.include_router(cust_router)
+        LOGGER.info("Customer routes registered via create_customers_router")
 
-        # Return updated workset
-        updated = state_db.get_workset(workset_id)
-        if updated is None:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Workset {workset_id} not found after update",
-            )
-        return WorksetResponse(**updated)
+        # File management routes (routes/files.py)
+        from daylib.routes.files import create_files_router, FileDependencies
+        file_deps = FileDependencies(customer_manager=customer_manager)
+        files_router = create_files_router(file_deps)
+        app.include_router(files_router)
+        LOGGER.info("File routes registered via create_files_router")
 
-    @app.post("/worksets/{workset_id}/lock", tags=["worksets"])
-    async def acquire_workset_lock(workset_id: str, owner_id: str = Query(..., description="Lock owner ID")):
-        """Acquire lock on a workset."""
-        success = state_db.acquire_lock(workset_id, owner_id)
-        if not success:
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail=f"Failed to acquire lock on workset {workset_id}",
-            )
-
-        return {"status": "locked", "workset_id": workset_id, "owner_id": owner_id}
-
-    @app.delete("/worksets/{workset_id}/lock", tags=["worksets"])
-    async def release_workset_lock(workset_id: str, owner_id: str = Query(..., description="Lock owner ID")):
-        """Release lock on a workset."""
-        success = state_db.release_lock(workset_id, owner_id)
-        if not success:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail=f"Failed to release lock on workset {workset_id} (not owner)",
-            )
-
-        return {"status": "unlocked", "workset_id": workset_id}
-
-    @app.get("/queue/stats", response_model=QueueStats, tags=["monitoring"])
-    async def get_queue_stats():
-        """Get queue statistics."""
-        queue_depth = state_db.get_queue_depth()
-
-        return QueueStats(
-            queue_depth=queue_depth,
-            total_worksets=sum(queue_depth.values()),
-            ready_worksets=queue_depth.get(WorksetState.READY.value, 0),
-            in_progress_worksets=queue_depth.get(WorksetState.IN_PROGRESS.value, 0),
-            error_worksets=queue_depth.get(WorksetState.ERROR.value, 0),
+        # Manifest routes (routes/manifests.py) - registry may be None (routes return 503)
+        from daylib.routes.manifests import create_manifests_router, ManifestDependencies
+        manifest_deps = ManifestDependencies(
+            customer_manager=customer_manager,
+            manifest_registry=manifest_registry,
         )
+        manifest_router = create_manifests_router(manifest_deps)
+        app.include_router(manifest_router)
+        LOGGER.info("Manifest routes registered via create_manifests_router")
 
-    @app.get("/scheduler/stats", response_model=SchedulingStats, tags=["monitoring"])
-    async def get_scheduler_stats():
-        """Get scheduler statistics."""
-        if not scheduler:
-            raise HTTPException(
-                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                detail="Scheduler not configured",
-            )
-
-        stats = scheduler.get_scheduling_stats()
-        return SchedulingStats(**stats)
-
-    @app.get("/worksets/next", response_model=Optional[WorksetResponse], tags=["scheduling"])
-    async def get_next_workset():
-        """Get the next workset to execute based on priority."""
-        if not scheduler:
-            # Fallback to simple priority-based selection
-            worksets = state_db.get_ready_worksets_prioritized(limit=1)
-            if not worksets:
-                return None
-            return WorksetResponse(**worksets[0])
-
-        next_workset = scheduler.get_next_workset()
-        if not next_workset:
-            return None
-
-        return WorksetResponse(**next_workset)
-
-    # ========== Cost Estimation Endpoints ==========
+    # ========== Cost Estimation Endpoint ==========
 
     @app.post("/api/estimate-cost", tags=["utilities"])
     async def estimate_workset_cost(
@@ -655,77 +495,27 @@ def create_app(
         priority: str = Body("normal", embed=True),
         data_size_gb: float = Body(0.0, embed=True),
     ):
-        """Estimate cost for a workset based on parameters.
-
-        Uses pipeline type, sample count, and coverage to estimate:
-        - vCPU hours required
-        - Estimated duration
-        - Cost in USD (based on current spot pricing)
-
-        Note: These are estimates. Actual costs depend on data complexity,
-        spot market conditions, and cluster utilization.
-        """
-        # Base vCPU-hours per sample by pipeline type
+        """Estimate cost for a workset based on parameters."""
         base_vcpu_hours_per_sample = {
-            "germline": 4.0,
-            "somatic": 8.0,
-            "rnaseq": 2.0,
-            "wgs": 12.0,
-            "wes": 3.0,
+            "germline": 4.0, "somatic": 8.0, "rnaseq": 2.0, "wgs": 12.0, "wes": 3.0,
         }
-
         base_hours = base_vcpu_hours_per_sample.get(pipeline_type, 4.0)
-
-        # Adjust for coverage (30x is baseline)
         coverage_factor = estimated_coverage / 30.0
-
-        # Calculate total vCPU hours
         vcpu_hours = base_hours * sample_count * coverage_factor
-
-        # Estimate duration assuming 16 vCPU instance average
         avg_vcpus = 16
         duration_hours = vcpu_hours / avg_vcpus
-
-        # Base cost per vCPU-hour (typical spot pricing)
-        cost_per_vcpu_hour = {
-            "urgent": 0.08,  # On-demand pricing
-            "high": 0.08,
-            "normal": 0.03,  # Spot pricing
-            "low": 0.015,    # Interruptible spot
-        }
-
+        cost_per_vcpu_hour = {"urgent": 0.08, "high": 0.08, "normal": 0.03, "low": 0.015}
         base_cost = cost_per_vcpu_hour.get(priority, 0.03)
-
-        # Calculate compute cost
         compute_cost = vcpu_hours * base_cost
-
-        # Storage cost estimate ($0.023/GB-month, estimate 1 week)
-        # Data size = sample_count * 50GB average per sample if not provided
         if data_size_gb <= 0:
             data_size_gb = sample_count * 50.0
-        storage_cost = data_size_gb * 0.023 / 4  # ~1 week
-
-        # FSx Lustre cost (if applicable) - $0.14/GB-month
-        fsx_cost = data_size_gb * 0.14 / 4  # ~1 week
-
-        # Data transfer cost (estimate 10% of data out at $0.09/GB)
+        storage_cost = data_size_gb * 0.023 / 4
+        fsx_cost = data_size_gb * 0.14 / 4
         transfer_cost = data_size_gb * 0.10 * 0.09
-
-        # Apply efficiency formula to storage costs
-        # Formula: total_size / (total_size - (total_size * 0.98))
         efficiency_multiplier = _calculate_cost_with_efficiency(data_size_gb)
         adjusted_storage_cost = storage_cost * efficiency_multiplier if efficiency_multiplier > 0 else storage_cost
-
-        # Total estimated cost
         total_cost = compute_cost + adjusted_storage_cost + fsx_cost + transfer_cost
-
-        # Priority multipliers
-        priority_multiplier = {
-            "urgent": 2.0,
-            "high": 1.5,
-            "normal": 1.0,
-            "low": 0.6,
-        }
+        priority_multiplier = {"urgent": 2.0, "high": 1.5, "normal": 1.0, "low": 0.6}
         multiplier = priority_multiplier.get(priority, 1.0)
 
         return {
@@ -755,843 +545,6 @@ def create_app(
             ],
         }
 
-    # ========== Customer Management Endpoints ==========
-
-    if customer_manager:
-        @app.post("/customers", response_model=CustomerResponse, tags=["customers"])
-        async def create_customer(
-            customer: CustomerCreate,
-            current_user: Optional[Dict] = Depends(get_current_user),
-        ):
-            """Create a new customer with provisioned resources."""
-            config = customer_manager.onboard_customer(
-                customer_name=customer.customer_name,
-                email=customer.email,
-                max_concurrent_worksets=customer.max_concurrent_worksets,
-                max_storage_gb=customer.max_storage_gb,
-                billing_account_id=customer.billing_account_id,
-                cost_center=customer.cost_center,
-            )
-
-            return CustomerResponse(
-                customer_id=config.customer_id,
-                customer_name=config.customer_name,
-                email=config.email,
-                s3_bucket=config.s3_bucket,
-                max_concurrent_worksets=config.max_concurrent_worksets,
-                max_storage_gb=config.max_storage_gb,
-                billing_account_id=config.billing_account_id,
-                cost_center=config.cost_center,
-            )
-
-        @app.get("/customers/{customer_id}", response_model=CustomerResponse, tags=["customers"])
-        async def get_customer(
-            customer_id: str,
-            current_user: Optional[Dict] = Depends(get_current_user),
-        ):
-            """Get customer details."""
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            return CustomerResponse(
-                customer_id=config.customer_id,
-                customer_name=config.customer_name,
-                email=config.email,
-                s3_bucket=config.s3_bucket,
-                max_concurrent_worksets=config.max_concurrent_worksets,
-                max_storage_gb=config.max_storage_gb,
-                billing_account_id=config.billing_account_id,
-                cost_center=config.cost_center,
-            )
-
-        @app.patch("/api/v1/customers/{customer_id}", response_model=CustomerResponse, tags=["customers"])
-        async def update_customer(
-            customer_id: str,
-            update: CustomerUpdate,
-            request: Request,
-            current_user: Optional[Dict] = Depends(get_current_user),
-        ):
-            """Update customer details (partial update).
-
-            Customers can only update their own account.
-            Admins can update any account.
-            """
-            # Get session customer for authorization check
-            session_customer_id = request.session.get("customer_id") if hasattr(request, "session") else None
-            session_email = request.session.get("user_email") if hasattr(request, "session") else None
-
-            # Check if user is admin
-            is_admin = False
-            if session_email:
-                session_customer = customer_manager.get_customer_by_email(session_email)
-                if session_customer and session_customer.is_admin:
-                    is_admin = True
-
-            # Authorization: must be own account or admin
-            if not is_admin and session_customer_id != customer_id:
-                raise HTTPException(
-                    status_code=status.HTTP_403_FORBIDDEN,
-                    detail="You can only update your own account",
-                )
-
-            # Perform the update
-            config = customer_manager.update_customer(
-                customer_id=customer_id,
-                customer_name=update.customer_name,
-                email=update.email,
-                billing_account_id=update.billing_account_id,
-                cost_center=update.cost_center,
-                max_concurrent_worksets=update.max_concurrent_worksets,
-                max_storage_gb=update.max_storage_gb,
-            )
-
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            return CustomerResponse(
-                customer_id=config.customer_id,
-                customer_name=config.customer_name,
-                email=config.email,
-                s3_bucket=config.s3_bucket,
-                max_concurrent_worksets=config.max_concurrent_worksets,
-                max_storage_gb=config.max_storage_gb,
-                billing_account_id=config.billing_account_id,
-                cost_center=config.cost_center,
-            )
-
-        @app.get("/customers", response_model=List[CustomerResponse], tags=["customers"])
-        async def list_customers(
-            current_user: Optional[Dict] = Depends(get_current_user),
-        ):
-            """List all customers."""
-            configs = customer_manager.list_customers()
-
-            return [
-                CustomerResponse(
-                    customer_id=c.customer_id,
-                    customer_name=c.customer_name,
-                    email=c.email,
-                    s3_bucket=c.s3_bucket,
-                    max_concurrent_worksets=c.max_concurrent_worksets,
-                    max_storage_gb=c.max_storage_gb,
-                    billing_account_id=c.billing_account_id,
-                    cost_center=c.cost_center,
-                )
-                for c in configs
-            ]
-
-        @app.get("/customers/{customer_id}/usage", tags=["customers"])
-        async def get_customer_usage(
-            customer_id: str,
-            current_user: Optional[Dict] = Depends(get_current_user),
-        ):
-            """Get customer resource usage."""
-            usage = customer_manager.get_customer_usage(customer_id)
-            if not usage:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            return usage
-
-        # ========== File Management Endpoints ==========
-
-        @app.get("/api/customers/{customer_id}/files", tags=["files"])
-        async def list_customer_files(
-            customer_id: str,
-            prefix: str = "",
-        ):
-            """List files in customer's S3 bucket.
-
-            Note: This endpoint does not require authentication to support portal usage.
-            In production, you may want to add authentication or rate limiting.
-            """
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            try:
-                s3 = boto3.client("s3")
-                response = s3.list_objects_v2(
-                    Bucket=config.s3_bucket,
-                    Prefix=prefix,
-                    Delimiter="/",
-                )
-
-                files = []
-
-                # Add folders (CommonPrefixes)
-                for cp in response.get("CommonPrefixes", []):
-                    folder_path = cp["Prefix"]
-                    folder_name = folder_path.rstrip("/").split("/")[-1]
-                    files.append({
-                        "key": folder_path,
-                        "name": folder_name,
-                        "type": "folder",
-                        "size": 0,
-                        "size_formatted": "-",
-                        "modified": None,
-                        "icon": "folder",
-                    })
-
-                # Add files (Contents)
-                for obj in response.get("Contents", []):
-                    key = obj["Key"]
-                    # Skip the prefix itself
-                    if key == prefix:
-                        continue
-                    name = key.split("/")[-1]
-                    if not name:
-                        continue
-                    size = obj["Size"]
-                    files.append({
-                        "key": key,
-                        "name": name,
-                        "type": "file",
-                        "size": size,
-                        "size_formatted": _format_file_size(size),
-                        "modified": obj["LastModified"].isoformat() if obj.get("LastModified") else None,
-                        "icon": _get_file_icon(name),
-                    })
-
-                return {"files": files, "prefix": prefix, "bucket": config.s3_bucket}
-
-            except Exception as e:
-                LOGGER.error("Failed to list files: %s", str(e))
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail=str(e),
-                )
-
-        @app.post("/api/customers/{customer_id}/files/upload", tags=["files"])
-        async def upload_file(
-            customer_id: str,
-            file: UploadFile = File(...),
-            prefix: str = Form(""),
-        ):
-            """Upload a file to customer's S3 bucket.
-
-            This endpoint proxies the upload through the server to avoid S3 CORS issues.
-            For large files in production, consider configuring S3 CORS and using presigned URLs.
-            """
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            try:
-                s3 = boto3.client("s3")
-                # Build the full key with prefix
-                key = f"{prefix}{file.filename}" if prefix else file.filename
-
-                # Read file content and upload to S3
-                content = await file.read()
-                s3.put_object(
-                    Bucket=config.s3_bucket,
-                    Key=key,
-                    Body=content,
-                    ContentType=file.content_type or "application/octet-stream",
-                )
-
-                LOGGER.info(f"Uploaded {key} to bucket {config.s3_bucket}")
-                return {"success": True, "key": key, "bucket": config.s3_bucket}
-
-            except Exception as e:
-                LOGGER.error("Failed to upload file: %s", str(e))
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail=str(e),
-                )
-
-        @app.post("/api/customers/{customer_id}/files/create-folder", tags=["files"])
-        async def create_folder(
-            customer_id: str,
-            folder_path: str = Body(..., embed=True),
-        ):
-            """Create a folder in customer's S3 bucket.
-
-            Note: This endpoint does not require authentication to support portal usage.
-            In production, you may want to add authentication or rate limiting.
-            """
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            try:
-                s3 = boto3.client("s3")
-                # S3 folders are just empty objects with trailing slash
-                folder_key = folder_path.rstrip("/") + "/"
-                s3.put_object(Bucket=config.s3_bucket, Key=folder_key, Body=b"")
-
-                # Also create a .hold file to prevent the folder from disappearing
-                # (S3 doesn't truly have folders, so an empty folder marker can disappear)
-                hold_file_key = folder_key.rstrip("/") + "/.hold"
-                s3.put_object(Bucket=config.s3_bucket, Key=hold_file_key, Body=b"")
-
-                LOGGER.info(f"Created folder {folder_key} in bucket {config.s3_bucket} (with .hold file)")
-                return {"success": True, "folder": folder_key}
-
-            except Exception as e:
-                LOGGER.error("Failed to create folder: %s", str(e))
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail=str(e),
-                )
-
-        @app.get("/api/customers/{customer_id}/files/{file_key:path}/preview", tags=["files"])
-        async def preview_file(
-            customer_id: str,
-            file_key: str,
-            lines: int = 20,
-        ):
-            """Preview file contents.
-
-            For compressed files (.gz, .tgz, .tar.gz), decompresses and shows first N lines.
-            For text files, shows first N lines directly.
-            For binary files, returns a message indicating preview is not available.
-            """
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            try:
-                s3 = boto3.client("s3")
-
-                # Get file metadata first
-                head = s3.head_object(Bucket=config.s3_bucket, Key=file_key)
-                file_size = head.get("ContentLength", 0)
-                content_type = head.get("ContentType", "application/octet-stream")
-
-                # Determine file type from extension
-                file_lower = file_key.lower()
-                is_gzip = file_lower.endswith(".gz") or file_lower.endswith(".gzip")
-                is_tar_gz = file_lower.endswith(".tar.gz") or file_lower.endswith(".tgz")
-                is_zip = file_lower.endswith(".zip")
-
-                # Text-like extensions
-                text_extensions = {
-                    ".txt", ".log", ".csv", ".tsv", ".json", ".xml", ".html", ".htm",
-                    ".yaml", ".yml", ".md", ".rst", ".py", ".js", ".ts", ".sh", ".bash",
-                    ".r", ".R", ".pl", ".rb", ".java", ".c", ".cpp", ".h", ".hpp",
-                    ".fastq", ".fq", ".fasta", ".fa", ".sam", ".vcf", ".bed", ".gff", ".gtf",
-                }
-
-                # Check if it's a text file (or compressed text)
-                base_name = file_key
-                if is_gzip and not is_tar_gz:
-                    base_name = file_key[:-3] if file_lower.endswith(".gz") else file_key[:-5]
-
-                ext = "." + base_name.split(".")[-1] if "." in base_name else ""
-                is_text = ext.lower() in text_extensions or content_type.startswith("text/")
-
-                # For very large files, limit how much we download
-                max_download = 10 * 1024 * 1024  # 10MB max to download for preview
-
-                # Get file content (limited range for large files)
-                if file_size > max_download:
-                    response = s3.get_object(
-                        Bucket=config.s3_bucket,
-                        Key=file_key,
-                        Range=f"bytes=0-{max_download}"
-                    )
-                else:
-                    response = s3.get_object(Bucket=config.s3_bucket, Key=file_key)
-
-                body = response["Body"].read()
-                preview_lines = []
-                file_type = "text"
-
-                if is_tar_gz:
-                    # Handle .tar.gz or .tgz - list contents and show first file preview
-                    file_type = "tar.gz"
-                    try:
-                        with tarfile.open(fileobj=io.BytesIO(body), mode="r:gz") as tar:
-                            members = tar.getnames()[:20]  # First 20 entries
-                            preview_lines.append(f"=== Archive contents ({len(tar.getnames())} files) ===")
-                            for m in members:
-                                preview_lines.append(m)
-                            if len(tar.getnames()) > 20:
-                                preview_lines.append(f"... and {len(tar.getnames()) - 20} more files")
-                    except Exception as e:
-                        preview_lines = [f"Error reading tar.gz: {str(e)}"]
-
-                elif is_gzip:
-                    # Handle .gz files - decompress and show content
-                    file_type = "gzip"
-                    try:
-                        decompressed = gzip.decompress(body)
-                        text = decompressed.decode("utf-8", errors="replace")
-                        preview_lines = text.split("\n")[:lines]
-                    except Exception as e:
-                        preview_lines = [f"Error decompressing: {str(e)}"]
-
-                elif is_zip:
-                    # Handle .zip files - list contents
-                    file_type = "zip"
-                    try:
-                        with zipfile.ZipFile(io.BytesIO(body)) as zf:
-                            names = zf.namelist()[:20]
-                            preview_lines.append(f"=== Archive contents ({len(zf.namelist())} files) ===")
-                            for name in names:
-                                preview_lines.append(name)
-                            if len(zf.namelist()) > 20:
-                                preview_lines.append(f"... and {len(zf.namelist()) - 20} more files")
-                    except Exception as e:
-                        preview_lines = [f"Error reading zip: {str(e)}"]
-
-                elif is_text or file_size < 1024 * 1024:  # Try text for small files
-                    # Try to decode as text
-                    try:
-                        text = body.decode("utf-8", errors="replace")
-                        preview_lines = text.split("\n")[:lines]
-                    except Exception:
-                        file_type = "binary"
-                        preview_lines = ["[Binary file - preview not available]"]
-                else:
-                    file_type = "binary"
-                    preview_lines = ["[Binary file - preview not available]"]
-
-                return {
-                    "filename": file_key.split("/")[-1],
-                    "file_type": file_type,
-                    "size": file_size,
-                    "lines": preview_lines,
-                    "total_lines": len(preview_lines),
-                    "truncated": len(preview_lines) >= lines,
-                }
-
-            except s3.exceptions.NoSuchKey:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"File not found: {file_key}",
-                )
-            except Exception as e:
-                LOGGER.error("Failed to preview file: %s", str(e))
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail=str(e),
-                )
-
-        @app.get("/api/customers/{customer_id}/files/{file_key:path}/download-url", tags=["files"])
-        async def get_download_url(
-            customer_id: str,
-            file_key: str,
-        ):
-            """Get presigned URL for file download.
-
-            Note: This endpoint does not require authentication to support portal usage.
-            In production, you may want to add authentication or rate limiting.
-            """
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            try:
-                s3 = boto3.client("s3")
-                url = s3.generate_presigned_url(
-                    "get_object",
-                    Params={"Bucket": config.s3_bucket, "Key": file_key},
-                    ExpiresIn=3600,
-                )
-                return {"url": url}
-
-            except Exception as e:
-                LOGGER.error("Failed to generate download URL: %s", str(e))
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail=str(e),
-                )
-
-        @app.delete("/api/customers/{customer_id}/files/{file_key:path}", tags=["files"])
-        async def delete_file(
-            customer_id: str,
-            file_key: str,
-        ):
-            """Delete a file from customer's S3 bucket.
-
-            Note: This endpoint does not require authentication to support portal usage.
-            In production, you may want to add authentication or rate limiting.
-            """
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            try:
-                s3 = boto3.client("s3")
-                s3.delete_object(Bucket=config.s3_bucket, Key=file_key)
-                return {"success": True, "deleted": file_key}
-
-            except Exception as e:
-                LOGGER.error("Failed to delete file: %s", str(e))
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail=str(e),
-                )
-
-
-        # ========== Customer Manifest Endpoints ==========
-
-        @app.get("/api/customers/{customer_id}/manifests", tags=["customer-manifests"])
-        async def list_customer_manifests(
-            customer_id: str,
-            limit: int = Query(200, ge=1, le=500),
-        ):
-            """List saved manifests for a customer (metadata only)."""
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-            if not manifest_registry:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Manifest storage not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            manifests = manifest_registry.list_customer_manifests(customer_id, limit=limit)
-            return {"manifests": manifests}
-
-        @app.post(
-            "/api/customers/{customer_id}/manifests",
-            tags=["customer-manifests"],
-            status_code=status.HTTP_201_CREATED,
-        )
-        async def save_customer_manifest(
-            customer_id: str,
-            tsv_content: str = Body(..., embed=True),
-            name: Optional[str] = Body(None, embed=True),
-            description: Optional[str] = Body(None, embed=True),
-        ):
-            """Save a stage_samples.tsv manifest for later reuse."""
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-            if not manifest_registry:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Manifest storage not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            try:
-                saved = manifest_registry.save_manifest(
-                    customer_id=customer_id,
-                    tsv_content=tsv_content,
-                    name=name,
-                    description=description,
-                )
-            except ManifestTooLargeError as e:
-                raise HTTPException(status_code=413, detail=str(e))
-            except ValueError as e:
-                raise HTTPException(status_code=400, detail=str(e))
-
-            download_url = f"/api/customers/{customer_id}/manifests/{saved.manifest_id}/download"
-            return {
-                "manifest": saved.to_metadata_dict(),
-                "download_url": download_url,
-            }
-
-        @app.get("/api/customers/{customer_id}/manifests/{manifest_id}", tags=["customer-manifests"])
-        async def get_customer_manifest_metadata(customer_id: str, manifest_id: str):
-            """Get saved manifest metadata (not content)."""
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-            if not manifest_registry:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Manifest storage not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            m = manifest_registry.get_manifest(customer_id=customer_id, manifest_id=manifest_id)
-            if not m:
-                raise HTTPException(status_code=404, detail="Manifest not found")
-            return {"manifest": m.to_metadata_dict()}
-
-        @app.get(
-            "/api/customers/{customer_id}/manifests/{manifest_id}/download",
-            tags=["customer-manifests"],
-        )
-        async def download_customer_manifest(customer_id: str, manifest_id: str):
-            """Download the saved stage_samples.tsv content."""
-            if not customer_manager:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Customer management not configured",
-                )
-            if not manifest_registry:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail="Manifest storage not configured",
-                )
-
-            config = customer_manager.get_customer_config(customer_id)
-            if not config:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Customer {customer_id} not found",
-                )
-
-            tsv = manifest_registry.get_manifest_tsv(customer_id=customer_id, manifest_id=manifest_id)
-            if tsv is None:
-                raise HTTPException(status_code=404, detail="Manifest not found")
-
-            filename = f"{manifest_id}.stage_samples.tsv"
-            return Response(
-                content=tsv,
-                media_type="text/tab-separated-values",
-                headers={"Content-Disposition": f"attachment; filename=\"{filename}\""},
-            )
-
-        # NOTE: Customer workset routes (list, get, create, cancel, retry, archive, delete) moved to daylib/routes/customer_worksets.py
-
-        @app.get("/api/admin/worksets/{workset_id}/command-log", tags=["admin"])
-        async def get_workset_command_log(
-            request: Request,
-            workset_id: str,
-            grep: Optional[str] = Query(None, description="Filter log entries containing this text"),
-            label: Optional[str] = Query(None, description="Filter by command label (e.g., stage_samples, clone_pipeline)"),
-            limit: int = Query(1000, ge=1, le=10000, description="Maximum lines to return"),
-        ):
-            """Admin-only endpoint to view command logs for a workset.
-
-            Retrieves the workset_command.log file from S3 which contains all SSH
-            and AWS CLI commands executed by the monitor for this workset.
-
-            Log format includes:
-            - Timestamp
-            - Label (e.g., stage_samples, clone_pipeline, export_results)
-            - Command executed
-            - Exit code
-            - Stdout (truncated)
-            - Stderr (truncated)
-
-            Args:
-                workset_id: Workset identifier
-                grep: Optional text filter - only return entries containing this text
-                label: Optional label filter - only return entries with this command label
-                limit: Maximum lines to return (default 1000)
-
-            Returns:
-                JSON with log entries and metadata
-            """
-            # Admin-only authorization
-            is_admin = getattr(request, "session", {}).get("is_admin", False)
-            if not is_admin:
-                raise HTTPException(
-                    status_code=status.HTTP_403_FORBIDDEN,
-                    detail="Admin access required to view command logs",
-                )
-
-            workset = state_db.get_workset(workset_id)
-            if not workset:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"Workset {workset_id} not found",
-                )
-
-            bucket = workset.get("bucket")
-            prefix = workset.get("prefix", "").rstrip("/") + "/"
-            if not bucket or not prefix:
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail="Workset missing bucket or prefix",
-                )
-
-            # Build the S3 key for the command log
-            command_log_key = f"{prefix}workset_command.log"
-
-            try:
-                # Create boto3 session using settings
-                app_settings = request.app.state.settings
-                session_kwargs = {"region_name": app_settings.get_effective_region()}
-                if app_settings.aws_profile:
-                    session_kwargs["profile_name"] = app_settings.aws_profile
-                session = boto3.Session(**session_kwargs)
-                s3_client = session.client("s3")
-
-                # Fetch the command log from S3
-                try:
-                    response = s3_client.get_object(Bucket=bucket, Key=command_log_key)
-                    log_content = response["Body"].read().decode("utf-8")
-                except s3_client.exceptions.NoSuchKey:
-                    return {
-                        "workset_id": workset_id,
-                        "log_available": False,
-                        "message": "No command log found for this workset",
-                        "entries": [],
-                        "entry_count": 0,
-                    }
-
-            except Exception as e:
-                LOGGER.warning("Failed to fetch command log for %s: %s", sanitize_for_log(workset_id), str(e))
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail=f"Failed to fetch command log: {str(e)}",
-                )
-
-            # Parse and filter log entries
-            lines = log_content.split("\n")
-            entries: List[str] = []
-            current_entry: List[str] = []
-            entry_count = 0
-
-            for line in lines:
-                # New entry starts with timestamp marker
-                if line.startswith("[") and "=====" in line:
-                    if current_entry:
-                        entry_text = "\n".join(current_entry)
-                        # Apply filters
-                        include_entry = True
-                        if grep and grep.lower() not in entry_text.lower():
-                            include_entry = False
-                        if label:
-                            label_match = f"LABEL: {label}"
-                            if label_match.lower() not in entry_text.lower():
-                                include_entry = False
-                        if include_entry:
-                            entries.append(entry_text)
-                            entry_count += 1
-                    current_entry = [line]
-                else:
-                    current_entry.append(line)
-
-            # Don't forget the last entry
-            if current_entry:
-                entry_text = "\n".join(current_entry)
-                include_entry = True
-                if grep and grep.lower() not in entry_text.lower():
-                    include_entry = False
-                if label:
-                    label_match = f"LABEL: {label}"
-                    if label_match.lower() not in entry_text.lower():
-                        include_entry = False
-                if include_entry:
-                    entries.append(entry_text)
-                    entry_count += 1
-
-            # Apply limit (on lines, not entries for consistency)
-            total_lines = sum(len(e.split("\n")) for e in entries)
-            if total_lines > limit:
-                # Truncate entries to fit within limit
-                limited_entries = []
-                line_count = 0
-                for entry in entries:
-                    entry_lines = len(entry.split("\n"))
-                    if line_count + entry_lines <= limit:
-                        limited_entries.append(entry)
-                        line_count += entry_lines
-                    else:
-                        break
-                entries = limited_entries
-
-            return {
-                "workset_id": workset_id,
-                "log_available": True,
-                "bucket": bucket,
-                "key": command_log_key,
-                "entry_count": entry_count,
-                "entries_returned": len(entries),
-                "filters_applied": {
-                    "grep": grep,
-                    "label": label,
-                    "limit": limit,
-                },
-                "entries": entries,
-            }
-
-        # NOTE: Customer workset routes (restore, logs, metrics, snakemake-log) moved to daylib/routes/customer_worksets.py
-
-        # NOTE: Dashboard routes moved to daylib/routes/dashboard.py
-        pass  # End of customer_manager block for routes that remain here
-
-    # NOTE: Billing routes moved to daylib/routes/billing.py
-
     # ========== Workset Validation Endpoints ==========
 
     if validator:
@@ -1604,7 +557,6 @@ def create_app(
         ):
             """Validate a workset configuration."""
             result = validator.validate_workset(bucket, prefix, dry_run)
-
             return WorksetValidationResponse(
                 is_valid=result.is_valid,
                 errors=result.errors,
@@ -1631,477 +583,8 @@ def create_app(
             "max_retries": request.max_retries,
             "estimated_coverage": request.estimated_coverage,
         }
-
         yaml_content = yaml.dump(work_config, default_flow_style=False, sort_keys=False)
-
-        return {
-            "yaml_content": yaml_content,
-            "config": work_config,
-        }
-
-    # ========== S3 Discovery Endpoint ==========
-
-    @app.post("/api/s3/discover-samples", tags=["utilities"])
-    async def discover_samples_from_s3(
-        request: Request,
-        bucket: str = Body(..., embed=True),
-        prefix: str = Body(..., embed=True),
-        current_user: Optional[Dict] = Depends(get_current_user),
-    ):
-        """Discover FASTQ samples from an S3 path.
-
-        Lists files in the given S3 location and automatically pairs R1/R2 files
-        into samples. Also attempts to parse daylily_work.yaml if present.
-        """
-        samples = []
-        yaml_content = None
-        files_found = []
-        all_keys_found = []  # For debugging
-
-        LOGGER.info("S3 Discovery: Starting discovery for bucket=%s, prefix=%s", bucket, prefix)
-
-        try:
-            # Create boto3 session using settings
-            app_settings = request.app.state.settings
-            session_kwargs = {"region_name": app_settings.get_effective_region()}
-            if app_settings.aws_profile:
-                session_kwargs["profile_name"] = app_settings.aws_profile
-            session = boto3.Session(**session_kwargs)
-            s3_client = session.client("s3")
-
-            # Normalize prefix - handle various input formats
-            # Strip whitespace and handle both with/without trailing slash
-            normalized_prefix = prefix.strip()
-            if normalized_prefix:
-                # Remove leading slash if present
-                normalized_prefix = normalized_prefix.lstrip("/")
-                # Ensure trailing slash for listing
-                if not normalized_prefix.endswith("/"):
-                    normalized_prefix += "/"
-
-            LOGGER.info("S3 Discovery: Using normalized prefix: '%s'", normalized_prefix)
-
-            # List objects in the S3 path
-            paginator = s3_client.get_paginator("list_objects_v2")
-            total_objects = 0
-
-            for page in paginator.paginate(Bucket=bucket, Prefix=normalized_prefix):
-                for obj in page.get("Contents", []):
-                    total_objects += 1
-                    key = obj["Key"]
-                    filename = key.split("/")[-1]
-                    all_keys_found.append(key)
-
-                    # Skip directory markers (empty keys ending with /)
-                    if not filename:
-                        continue
-
-                    # Check for daylily_work.yaml (case-insensitive)
-                    if filename.lower() == "daylily_work.yaml":
-                        LOGGER.info("S3 Discovery: Found daylily_work.yaml at %s", key)
-                        try:
-                            response = s3_client.get_object(Bucket=bucket, Key=key)
-                            yaml_content = response["Body"].read().decode("utf-8")
-                            LOGGER.info("S3 Discovery: Successfully read daylily_work.yaml")
-                        except Exception as e:
-                            LOGGER.warning("S3 Discovery: Failed to read daylily_work.yaml: %s", str(e))
-
-                    # Check for FASTQ files - extended patterns
-                    fastq_extensions = [".fastq", ".fq", ".fastq.gz", ".fq.gz", ".fastq.bz2", ".fq.bz2"]
-                    if any(filename.lower().endswith(ext) for ext in fastq_extensions):
-                        LOGGER.debug("S3 Discovery: Found FASTQ file: %s", filename)
-                        files_found.append({
-                            "key": key,
-                            "filename": filename,
-                            "size": obj.get("Size", 0),
-                        })
-
-            LOGGER.info("S3 Discovery: Found %d total objects, %d FASTQ files", total_objects, len(files_found))
-
-            # If we found a daylily_work.yaml, parse samples from it
-            if yaml_content:
-                try:
-                    yaml_data = yaml.safe_load(yaml_content)
-                    if yaml_data and isinstance(yaml_data.get("samples"), list):
-                        for sample in yaml_data["samples"]:
-                            if isinstance(sample, dict):
-                                samples.append({
-                                    "sample_id": sample.get("sample_id") or sample.get("id") or sample.get("name", "unknown"),
-                                    "r1_file": sample.get("r1_file") or sample.get("r1") or sample.get("fq1", ""),
-                                    "r2_file": sample.get("r2_file") or sample.get("r2") or sample.get("fq2", ""),
-                                    "status": "pending",
-                                })
-                        LOGGER.info("S3 Discovery: Parsed %d samples from YAML", len(samples))
-                except Exception as e:
-                    LOGGER.warning("S3 Discovery: Failed to parse daylily_work.yaml: %s", str(e))
-
-            # If no samples from YAML, try to pair FASTQ files
-            if not samples and files_found:
-                LOGGER.info("S3 Discovery: No YAML samples, attempting to pair %d FASTQ files", len(files_found))
-
-                # Pattern matching for R1/R2 pairs - more flexible patterns
-                # Supports: sample_R1.fastq.gz, sample.R1.fastq.gz, sample_1.fastq.gz,
-                #           sample_R1_001.fastq.gz, sample_S1_L001_R1_001.fastq.gz (Illumina)
-                r1_patterns = [
-                    # Standard patterns: sample_R1.fastq.gz, sample.R1.fastq.gz
-                    re.compile(r"^(.+?)[._](R1|r1)[._]?.*\.(fastq|fq)(\.gz|\.bz2)?$", re.IGNORECASE),
-                    # Numeric patterns: sample_1.fastq.gz
-                    re.compile(r"^(.+?)[._]1[._]?.*\.(fastq|fq)(\.gz|\.bz2)?$", re.IGNORECASE),
-                    # Illumina patterns: sample_S1_L001_R1_001.fastq.gz
-                    re.compile(r"^(.+?)_S\d+_L\d+_R1_\d+\.(fastq|fq)(\.gz|\.bz2)?$", re.IGNORECASE),
-                ]
-                r2_patterns = [
-                    re.compile(r"^(.+?)[._](R2|r2)[._]?.*\.(fastq|fq)(\.gz|\.bz2)?$", re.IGNORECASE),
-                    re.compile(r"^(.+?)[._]2[._]?.*\.(fastq|fq)(\.gz|\.bz2)?$", re.IGNORECASE),
-                    re.compile(r"^(.+?)_S\d+_L\d+_R2_\d+\.(fastq|fq)(\.gz|\.bz2)?$", re.IGNORECASE),
-                ]
-
-                r1_files = {}
-                r2_files = {}
-
-                for f in files_found:
-                    filename = f["filename"]
-                    matched = False
-
-                    # Try R1 patterns
-                    for pattern in r1_patterns:
-                        match = pattern.match(filename)
-                        if match:
-                            sample_name = match.group(1)
-                            r1_files[sample_name] = f["key"]
-                            LOGGER.debug("S3 Discovery: Matched R1 file %s -> sample %s", filename, sample_name)
-                            matched = True
-                            break
-
-                    if not matched:
-                        # Try R2 patterns
-                        for pattern in r2_patterns:
-                            match = pattern.match(filename)
-                            if match:
-                                sample_name = match.group(1)
-                                r2_files[sample_name] = f["key"]
-                                LOGGER.debug("S3 Discovery: Matched R2 file %s -> sample %s", filename, sample_name)
-                                break
-
-                # Pair R1 and R2 files
-                all_sample_names = set(r1_files.keys()) | set(r2_files.keys())
-                LOGGER.info("S3 Discovery: Found %d R1 files, %d R2 files, %d unique sample names",
-                           len(r1_files), len(r2_files), len(all_sample_names))
-
-                for sample_name in sorted(all_sample_names):
-                    samples.append({
-                        "sample_id": sample_name,
-                        "r1_file": r1_files.get(sample_name, ""),
-                        "r2_file": r2_files.get(sample_name, ""),
-                        "status": "pending",
-                    })
-
-            LOGGER.info("S3 Discovery: Returning %d samples, %d files found", len(samples), len(files_found))
-
-            return {
-                "samples": samples,
-                "yaml_content": yaml_content,
-                "files_found": len(files_found),
-                "bucket": bucket,
-                "prefix": prefix,
-                "normalized_prefix": normalized_prefix,
-                "total_objects_scanned": total_objects,
-            }
-
-        except s3_client.exceptions.NoSuchBucket if 's3_client' in dir() else Exception as e:
-            if "NoSuchBucket" in str(type(e).__name__):
-                LOGGER.error("S3 Discovery: Bucket '%s' does not exist", bucket)
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"S3 bucket '{bucket}' not found",
-                )
-            LOGGER.error("S3 Discovery: Failed to discover samples from S3: %s", str(e))
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail=f"Failed to discover samples: {str(e)}",
-            )
-        except Exception as e:
-            LOGGER.error("S3 Discovery: Failed to discover samples from S3: %s", str(e))
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail=f"Failed to discover samples: {str(e)}",
-            )
-
-    # ========== S3 Bucket Validation Endpoint ==========
-
-    @app.post("/api/s3/validate-bucket", tags=["utilities"])
-    async def validate_s3_bucket(
-        bucket: str = Body(..., embed=True),
-        current_user: Optional[Dict] = Depends(get_current_user),
-    ):
-        """Validate an S3 bucket for Daylily use.
-
-        Checks:
-        - Bucket exists and is accessible
-        - Read permissions (list and get objects)
-        - Write permissions (put objects to worksets/ prefix)
-
-        Returns validation result with setup instructions if needed.
-        """
-        from daylib.s3_bucket_validator import S3BucketValidator
-
-        try:
-            validator = S3BucketValidator(region=region, profile=profile)
-            result = validator.validate_bucket(bucket)
-
-            # Generate setup instructions if not fully configured
-            instructions = None
-            if not result.is_fully_configured:
-                instructions = validator.get_setup_instructions(
-                    bucket, result, daylily_account_id="108782052779"
-                )
-
-            return {
-                "bucket": bucket,
-                "valid": result.is_valid,
-                "fully_configured": result.is_fully_configured,
-                "exists": result.exists,
-                "accessible": result.accessible,
-                "can_read": result.can_read,
-                "can_write": result.can_write,
-                "can_list": result.can_list,
-                "region": result.region,
-                "errors": result.errors,
-                "warnings": result.warnings,
-                "setup_instructions": instructions,
-            }
-        except Exception as e:
-            LOGGER.error("S3 Validation: Failed to validate bucket '%s': %s", bucket, str(e))
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail=f"Failed to validate bucket: {str(e)}",
-            )
-
-    @app.get("/api/s3/iam-policy/{bucket_name}", tags=["utilities"])
-    async def get_iam_policy_for_bucket(
-        bucket_name: str,
-        read_only: bool = False,
-        current_user: Optional[Dict] = Depends(get_current_user),
-    ):
-        """Generate IAM policy for accessing a customer S3 bucket.
-
-        Args:
-            bucket_name: S3 bucket name
-            read_only: If True, generate read-only policy
-
-        Returns:
-            IAM policy document that can be attached to a role/user.
-        """
-        from daylib.s3_bucket_validator import S3BucketValidator
-
-        validator = S3BucketValidator(region=region, profile=profile)
-        policy = validator.generate_iam_policy_for_bucket(bucket_name, read_only=read_only)
-
-        return {
-            "bucket": bucket_name,
-            "read_only": read_only,
-            "policy": policy,
-        }
-
-    @app.get("/api/s3/bucket-policy/{bucket_name}", tags=["utilities"])
-    async def get_bucket_policy_for_daylily(
-        bucket_name: str,
-        daylily_account_id: str = "108782052779",
-        current_user: Optional[Dict] = Depends(get_current_user),
-    ):
-        """Generate S3 bucket policy for cross-account Daylily access.
-
-        Args:
-            bucket_name: Customer's S3 bucket name
-            daylily_account_id: Daylily service account ID
-
-        Returns:
-            S3 bucket policy document to apply to customer bucket.
-        """
-        from daylib.s3_bucket_validator import S3BucketValidator
-
-        validator = S3BucketValidator(region=region, profile=profile)
-        policy = validator.generate_customer_bucket_policy(bucket_name, daylily_account_id)
-
-        return {
-            "bucket": bucket_name,
-            "daylily_account_id": daylily_account_id,
-            "policy": policy,
-            "apply_command": f"aws s3api put-bucket-policy --bucket {bucket_name} --policy file://bucket-policy.json",
-        }
-
-    @app.get("/api/s3/bucket-region/{bucket_name}", tags=["utilities"])
-    async def get_bucket_region(
-        bucket_name: str,
-        current_user: Optional[Dict] = Depends(get_current_user),
-    ):
-        """Get the AWS region where an S3 bucket is located.
-
-        Args:
-            bucket_name: S3 bucket name
-
-        Returns:
-            Bucket region information.
-        """
-        import boto3
-        from botocore.exceptions import ClientError
-
-        try:
-            session_kwargs = {}
-            if profile:
-                session_kwargs["profile_name"] = profile
-            session = boto3.Session(**session_kwargs)
-            s3_client = session.client("s3")
-
-            # get_bucket_location returns None for us-east-1, otherwise the region
-            response = s3_client.get_bucket_location(Bucket=bucket_name)
-            location = response.get("LocationConstraint")
-            # us-east-1 returns None or empty string
-            bucket_region = location if location else "us-east-1"
-
-            return {
-                "bucket": bucket_name,
-                "region": bucket_region,
-            }
-        except ClientError as e:
-            error_code = e.response.get("Error", {}).get("Code", "Unknown")
-            if error_code == "NoSuchBucket":
-                raise HTTPException(status_code=404, detail=f"Bucket '{bucket_name}' not found")
-            elif error_code == "AccessDenied":
-                raise HTTPException(status_code=403, detail=f"Access denied to bucket '{bucket_name}'")
-            else:
-                raise HTTPException(status_code=500, detail=f"Failed to get bucket region: {str(e)}")
-        except Exception as e:
-            raise HTTPException(status_code=500, detail=f"Failed to get bucket region: {str(e)}")
-
-    # ========== Cluster Management Endpoints ==========
-
-    @app.get("/api/clusters", tags=["clusters"])
-    async def list_clusters(
-        request: Request,
-        refresh: bool = Query(False, description="Force refresh cluster cache"),
-        fetch_status: bool = Query(False, description="Fetch SSH-based status (budget/jobs) for running headnodes"),
-    ):
-        """List all ParallelCluster instances across configured regions.
-
-        Returns cluster information including status, head node details,
-        compute fleet status, and relevant tags.
-
-        Uses caching to reduce API calls (5 minute TTL by default).
-        Set refresh=true to force a cache refresh.
-        Set fetch_status=true to also fetch budget and job queue info via SSH.
-        """
-        try:
-            from daylib.cluster_service import get_cluster_service
-            from daylib.ursa_config import get_ursa_config
-
-            # Use UrsaConfig for regions (preferred) with fallback to legacy env var
-            ursa_config = get_ursa_config()
-            if ursa_config.is_configured:
-                allowed_regions = ursa_config.get_allowed_regions()
-            else:
-                allowed_regions = settings.get_allowed_regions()
-
-            if not allowed_regions:
-                return {
-                    "clusters": [],
-                    "regions": [],
-                    "error": "No regions configured. Create ~/.ursa/ursa-config.yaml with region definitions.",
-                }
-
-            # Use global singleton to share cache across requests
-            service = get_cluster_service(
-                regions=allowed_regions,
-                aws_profile=ursa_config.aws_profile or settings.aws_profile,
-                cache_ttl_seconds=300,
-            )
-
-            all_clusters = service.get_all_clusters_with_status(
-                force_refresh=refresh,
-                fetch_ssh_status=fetch_status,
-            )
-            clusters_dicts = [c.to_dict() for c in all_clusters]
-
-            return {
-                "clusters": clusters_dicts,
-                "regions": allowed_regions,
-                "total_count": len(clusters_dicts),
-                "cached": not refresh,
-                "status_fetched": fetch_status,
-            }
-        except Exception as e:
-            LOGGER.error(f"Failed to list clusters: {e}")
-            return {
-                "clusters": [],
-                "regions": [],
-                "error": str(e),
-            }
-
-    @app.delete("/api/clusters/{cluster_name}", tags=["clusters"])
-    async def delete_cluster(
-        cluster_name: str,
-        region: str = Query(..., description="AWS region where the cluster is located"),
-        current_user: Optional[Dict] = Depends(get_current_user),
-    ):
-        """Delete a ParallelCluster instance (admin-only).
-
-        This is a destructive operation.
-        """
-        from daylib.file_api import _enforce_admin_only
-
-        _enforce_admin_only(current_user, operation="delete clusters")
-
-        try:
-            from daylib.cluster_service import get_cluster_service
-            from daylib.ursa_config import get_ursa_config
-
-            ursa_config = get_ursa_config()
-            if ursa_config.is_configured:
-                allowed_regions = ursa_config.get_allowed_regions()
-            else:
-                allowed_regions = settings.get_allowed_regions()
-
-            if not allowed_regions:
-                raise HTTPException(
-                    status_code=400,
-                    detail="No regions configured. Create ~/.ursa/ursa-config.yaml with region definitions.",
-                )
-
-            if region not in allowed_regions:
-                raise HTTPException(
-                    status_code=400,
-                    detail=f"Region '{region}' is not configured for this deployment",
-                )
-
-            service = get_cluster_service(
-                regions=allowed_regions,
-                aws_profile=ursa_config.aws_profile or settings.aws_profile,
-                cache_ttl_seconds=300,
-            )
-
-            result = service.delete_cluster(cluster_name, region)
-            if "error" in result:
-                raise HTTPException(
-                    status_code=500,
-                    detail=f"Failed to delete cluster: {result['error']}",
-                )
-
-            cmd = f"pcluster delete-cluster --region {region} -n {cluster_name}"
-            profile = ursa_config.aws_profile or settings.aws_profile
-
-            return {
-                "success": True,
-                "cluster_name": cluster_name,
-                "region": region,
-                "pcluster_command": cmd,
-                "aws_profile": profile,
-                "message": f"Cluster '{cluster_name}' deletion initiated",
-            }
-        except HTTPException:
-            raise
-        except Exception as e:
-            LOGGER.error("Failed to delete cluster %s: %s", sanitize_for_log(cluster_name), e)
-            raise HTTPException(status_code=500, detail=str(e))
+        return {"yaml_content": yaml_content, "config": work_config}
 
     # ========== Customer Portal Routes ==========
